+++ MSAL/src/public/MSALAuthenticationSchemePop.h
@@ -29,14 +29,38 @@
 
 NS_ASSUME_NONNULL_BEGIN
 
+@class MSALExternalPoPKeyPair;
+
 @interface MSALAuthenticationSchemePop : NSObject<MSALAuthenticationSchemeProtocol>
 
 @property (nonatomic, readonly) MSALAuthScheme scheme;
+
+/**
+ Caller supplied key material used to bind the PoP token. When nil, MSAL uses its default
+ device bound, keychain backed key pair.
+ */
+@property (nonatomic, readonly, nullable) MSALExternalPoPKeyPair *externalKeyPair;
 
 - (instancetype)initWithHttpMethod:(MSALHttpMethod)httpMethod
                         requestUrl:(NSURL *)requestUrl
                              nonce:(nullable NSString *)nonce
               additionalParameters:(nullable NSDictionary *)additionalParameters;
+
+/**
+ Initializes a PoP authentication scheme that binds the token to caller supplied key material.
+
+ @param httpMethod           The HTTP method of the request the PoP token will be used with.
+ @param requestUrl           The URL of the request the PoP token will be used with.
+ @param nonce                An optional server provided nonce. When nil, MSAL generates one.
+ @param additionalParameters Additional parameters to include in the scheme.
+ @param externalKeyPair      Caller owned key material used to bind the PoP token. When nil,
+                             MSAL falls back to its default device bound key pair.
+ */
+- (instancetype)initWithHttpMethod:(MSALHttpMethod)httpMethod
+                        requestUrl:(NSURL *)requestUrl
+                             nonce:(nullable NSString *)nonce
+              additionalParameters:(nullable NSDictionary *)additionalParameters
+                   externalKeyPair:(nullable MSALExternalPoPKeyPair *)externalKeyPair NS_DESIGNATED_INITIALIZER;
 
 - (instancetype)init NS_UNAVAILABLE;
 + (instancetype)new NS_UNAVAILABLE;