+++ CHANGELOG.md
@@ -1,3 +1,6 @@
+## TBD:
+* Add support for caller-provided external key material for PoP token binding #2854
+
 ## [2.13.0]
 * Update IdentityCore submodule to pull in DI foundation (common core #1810 WPJ, #1838 hardening, #1809 throttling)
 * Migrating MSAL automation pipeline to ACES shared pool.